+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/notifications/handler/ChatNotificationHandler.kt
@@ -50,6 +50,9 @@ internal class ChatNotificationHandler(
     private val actionsProvider: (notificationId: Int, channel: Channel, message: Message) -> List<Action>,
     private val notificationBuilderTransformer:
     (NotificationCompat.Builder, ChatNotification) -> NotificationCompat.Builder,
+    private val currentUserProvider: () -> User? = {
+        ChatClient.instance().getCurrentUser() ?: ChatClient.instance().getStoredUser()
+    },
 ) : NotificationHandler {
 
     private val sharedPreferences: SharedPreferences by lazy {
@@ -85,26 +88,50 @@ internal class ChatNotificationHandler(
         showNotificationInternal(ChatNotification.MessageNew(channel, message))
     }
 
-    private fun showNotificationInternal(chatNotification: ChatNotification) {
-        when (chatNotification) {
-            is ChatNotification.MessageNew -> showMessageNewNotification(chatNotification)
-            is ChatNotification.NotificationReminderDue -> showReminderDueNotification(chatNotification)
+    private fun showNotificationInternal(notification: ChatNotification) {
+        when (notification) {
+            is ChatNotification.MessageNew -> showMessageNewNotification(notification)
+            is ChatNotification.MessageUpdated -> showMessageUpdatedNotification(notification)
+            is ChatNotification.ReactionNew -> showReactionNewNotification(notification)
+            is ChatNotification.NotificationReminderDue -> showReminderDueNotification(notification)
         }
     }
 
-    private fun showMessageNewNotification(chatNotification: ChatNotification.MessageNew) {
-        val (channel, message) = chatNotification
+    private fun showMessageNewNotification(notification: ChatNotification.MessageNew) {
+        val (channel, message) = notification
         val notificationId: Int = System.nanoTime().toInt()
         val notificationSummaryId = getNotificationGroupSummaryId(channel.type, channel.id)
         addNotificationId(notificationId, notificationSummaryId)
-        showNotification(
-            notificationId,
-            notificationBuilderTransformer(
-                buildNotification(notificationId, channel, message),
-                chatNotification,
-            ).build(),
-        )
+        val notification = notificationBuilderTransformer(
+            buildNotification(notificationId, channel, message),
+            notification,
+        ).build()
+        showNotification(notificationId, notification)
         showNotification(notificationSummaryId, buildNotificationGroupSummary(channel, message).build())
+    }
+
+    private fun showMessageUpdatedNotification(notification: ChatNotification.MessageUpdated) {
+        // Note: Handled the same as MessageNew - perhaps in future we want to differentiate them
+        val (channel, message) = notification
+        val notificationId: Int = System.nanoTime().toInt()
+        val notificationSummaryId = getNotificationGroupSummaryId(channel.type, channel.id)
+        addNotificationId(notificationId, notificationSummaryId)
+        val notification = notificationBuilderTransformer(
+            buildNotification(notificationId, channel, message),
+            notification,
+        ).build()
+        showNotification(notificationId, notification)
+        showNotification(notificationSummaryId, buildNotificationGroupSummary(channel, message).build())
+    }
+
+    private fun showReactionNewNotification(notification: ChatNotification.ReactionNew) {
+        val notificationId = "${notification.message.id}:${notification.reactionUserId}:${notification.type}".hashCode()
+        addNotificationIdWithoutSummary(notificationId)
+        val notification = notificationBuilderTransformer(
+            buildReactionNewNotification(notification),
+            notification,
+        ).build()
+        showNotification(notificationId, notification)
     }
 
     private fun showReminderDueNotification(chatNotification: ChatNotification.NotificationReminderDue) {
@@ -123,8 +150,7 @@ internal class ChatNotificationHandler(
         channel: Channel,
         message: Message,
     ): NotificationCompat.Builder {
-        val currentUser = ChatClient.instance().getCurrentUser()
-            ?: ChatClient.instance().getStoredUser()
+        val currentUser = currentUserProvider()
         return getNotificationBuilder(
             contentTitle = channel.getNotificationContentTitle(),
             contentText = notificationTextFormatter(currentUser, message),
@@ -136,9 +162,17 @@ internal class ChatNotificationHandler(
         }
     }
 
+    private fun buildReactionNewNotification(notification: ChatNotification.ReactionNew): NotificationCompat.Builder {
+        return getNotificationBuilder(
+            contentTitle = notification.title,
+            contentText = notification.body,
+            groupKey = null,
+            intent = getNewMessageIntent(message = notification.message, channel = notification.channel),
+        )
+    }
+
     private fun buildReminderDueNotification(channel: Channel, message: Message): NotificationCompat.Builder {
-        val currentUser = ChatClient.instance().getCurrentUser()
-            ?: ChatClient.instance().getStoredUser()
+        val currentUser = currentUserProvider()
         return getNotificationBuilder(
             contentTitle = context.getString(R.string.stream_chat_notification_reminder_due_title),
             contentText = notificationTextFormatter(currentUser, message),