+++ stream-chat-android-core/src/main/java/io/getstream/chat/android/client/PayloadValidator.kt
@@ -16,6 +16,7 @@
 
 package io.getstream.chat.android.client
 
+@Suppress("TooManyFunctions")
 public object PayloadValidator {
 
     private const val KEY_SENDER_V1 = "sender_server"
@@ -23,17 +24,22 @@ public object PayloadValidator {
     private const val KEY_VERSION = "version"
     private const val KEY_TYPE = "type"
     private const val KEY_CHANNEL_ID = "channel_id"
-    private const val KEY_MESSAGE_ID = "message_id"
     private const val KEY_CHANNEL_TYPE = "channel_type"
+    private const val KEY_CID = "cid"
+    private const val KEY_MESSAGE_ID = "message_id"
 
     private const val VALUE_STREAM_SENDER = "stream.chat"
-    private const val VALUE_NEW_MESSAGE_TYPE = "message.new"
+    private const val VALUE_MESSAGE_NEW = "message.new"
+    private const val VALUE_MESSAGE_UPDATED = "message.updated"
+    private const val VALUE_REACTION_NEW = "reaction.new"
     private const val VALUE_NOTIFICATION_REMINDER_DUE = "notification.reminder_due"
     private const val VALUE_V1 = "v1"
     private const val VALUE_V2 = "v2"
 
     private val supportedNotificationTypes = listOf(
-        VALUE_NEW_MESSAGE_TYPE,
+        VALUE_MESSAGE_NEW,
+        VALUE_MESSAGE_UPDATED,
+        VALUE_REACTION_NEW,
         VALUE_NOTIFICATION_REMINDER_DUE,
     )
 
@@ -103,19 +109,28 @@ public object PayloadValidator {
      * @return true if the payload contains all needed fields for a new message.
      */
     private fun isValidNewMessageV2(payload: Map<String, Any?>): Boolean =
-        payload[KEY_TYPE] == VALUE_NEW_MESSAGE_TYPE &&
+        payload[KEY_TYPE] == VALUE_MESSAGE_NEW &&
             !(payload[KEY_CHANNEL_ID] as? String).isNullOrBlank() &&
             !(payload[KEY_MESSAGE_ID] as? String).isNullOrBlank() &&
             !(payload[KEY_CHANNEL_TYPE] as? String).isNullOrBlank()
 
     private fun isValidTypeV2(payload: Map<String, Any?>): Boolean =
-        isValidType(payload) &&
-            !(payload[KEY_CHANNEL_ID] as? String).isNullOrBlank() &&
-            !(payload[KEY_MESSAGE_ID] as? String).isNullOrBlank() &&
-            !(payload[KEY_CHANNEL_TYPE] as? String).isNullOrBlank()
+        isValidType(payload) && hasCid(payload) && hasMessageId(payload)
 
     private fun isValidType(payload: Map<String, Any?>): Boolean {
         val type = payload[KEY_TYPE] as? String ?: return false
         return type in supportedNotificationTypes
+    }
+
+    private fun hasMessageId(payload: Map<String, Any?>): Boolean {
+        return !(payload[KEY_MESSAGE_ID] as? String).isNullOrBlank()
+    }
+
+    private fun hasCid(payload: Map<String, Any?>): Boolean {
+        return (
+            !(payload[KEY_CHANNEL_ID] as? String).isNullOrBlank() &&
+                !(payload[KEY_CHANNEL_TYPE] as? String).isNullOrBlank()
+            ) ||
+            !(payload[KEY_CID] as? String).isNullOrBlank()
     }
 }