+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/viewmodel/messages/MessageListViewModel.kt
@@ -237,7 +237,12 @@ public class MessageListViewModel(
             is Event.UnpinMessage -> messageListController.unpinMessage(event.message)
             is Event.GiphyActionSelected -> onGiphyActionSelected(event)
             is Event.RetryMessage -> messageListController.resendMessage(event.message)
-            is Event.MessageReaction -> onMessageReaction(event.message, event.reactionType)
+            is Event.MessageReaction -> onMessageReaction(
+                message = event.message,
+                reactionType = event.reactionType,
+                emojiCode = event.emojiCode,
+                skipPush = event.skipPush,
+            )
             is Event.MuteUser -> messageListController.muteUser(event.user)
             is Event.UnmuteUser -> messageListController.unmuteUser(event.user)
             is Event.UnbanUser -> messageListController.unbanUser(event.user.id)
@@ -443,14 +448,18 @@ public class MessageListViewModel(
      *
      * @param message The message the user is reacting to.
      * @param reactionType The exact reaction type.
+     * @param emojiCode Optional emoji to be shown in the push notification delivered for the reaction
+     * (instead of :type:).
+     * @param skipPush If set to "true", skips sending push notification when reacting to a message.
      */
-    private fun onMessageReaction(message: Message, reactionType: String) {
+    private fun onMessageReaction(message: Message, reactionType: String, emojiCode: String?, skipPush: Boolean) {
         val reaction = Reaction(
             messageId = message.id,
             type = reactionType,
             score = 1,
+            emojiCode = emojiCode,
         )
-        messageListController.reactToMessage(reaction, message)
+        messageListController.reactToMessage(reaction, message, skipPush)
     }
 
     /**
@@ -642,10 +651,14 @@ public class MessageListViewModel(
          *
          * @param message The message the user is reacting to
          * @param reactionType The reaction type.
+         * @param emojiCode The optional emoji code to be shown in the push notification text, for the API.
+         * @param skipPush If set to "true", skips sending push notification when reacting to a message.
          */
         public data class MessageReaction(
             val message: Message,
             val reactionType: String,
+            val emojiCode: String? = null,
+            val skipPush: Boolean = false,
         ) : Event()
 
         /**